Added a couple new functions to .permissions.inc
These functions do some general querying, which will be used by the
XACML functions to create the XACML files. I modified a couple of the
other functions so that the queries are all done through the new
functions.

Ideally we would eventually move all of the queries into a couple of
sepcialized query functions, and then have others manipulate the data.