Updated the module to support XACML
The module will now create an appropriate XACML policy DS when it
is being created and when it is changing state. The only problem
left to solve is what happens when users are changed while objects
exist. Currently the changes are not carried over into the POLICY
stream until the object changes state.