Further XACML integration.
Enabled batch permission changing on the admin screens
in the workflow module. If XACML is enabled and permissions are
changed then a batch is kicked off that will modify all the XACML
POLICY datastreams to reflect the new settings.

Added batch functionality to the enable/disable XACML
in workflow screen. This function could take a very long time if
there are a large number of objects in workflow. It will change
EVERY object in the workflow to add an XACML datastream or
to delete the XACML datastream based on the setting on the
admin page.

The admin page function has the possibility of failing if the
user that invokes the function doesn't have the proper permissions
to modify every object in workflow. I added a message to warn users
and possibly we should add more protection in the future.